requirements.txt
absl-py==1.0.0
astunparse==1.6.3
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
enum34==1.1.10
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
micromlgen==1.1.26
numpy==1.21.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyXB==1.2.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
six==1.15.0
sklearn==0.0
sklearn-pandas==2.2.0
sklearn-pmml==0.1.2
sklearn2pmml==0.77.0
SQLAlchemy==1.4.28
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-macos==2.7.0
tensorflow-metal==0.3.0
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
